Criar classe de excecao para cadastro inexistente

+++ src/main/java/br/com/zup/Cadastros/cadastro/CadastroService.java
@@ -38,7 +38,11 @@ public class CadastroService {
     }
 
     public void deletarCadastro(String cpf) {
+        if(cadastroRepository.existsById(cpf)){
         cadastroRepository.deleteById(cpf);
+        } else {
+            throw new CadastroNaoExisteException();
+        }
     }
 
 }
